split datastore & transactor types into separate objects
