+++ transactor/src/main/scala/io/mediachain/types/CborSerialization.scala
@@ -1,4 +1,4 @@
-package io.mediachain.transactor
+package io.mediachain.types
 
 import io.mediachain.multihash.MultiHash
 
@@ -6,8 +6,7 @@ import io.mediachain.multihash.MultiHash
 object CborSerialization {
 
   import cats.data.Xor
-  import io.mediachain.transactor.Dummies.DummyReference
-  import io.mediachain.transactor.Types._
+  import io.mediachain.types.DatastoreTypes._
   import io.mediachain.util.cbor.CborAST._
   import io.mediachain.util.cbor.CborCodec
