+++ transactor/src/main/scala/io/mediachain/types/Datastore.scala
@@ -1,0 +1,8 @@
+package io.mediachain.types
+
+// Datastore interface
+trait Datastore {
+  import DatastoreTypes.{DataObject, Reference}
+
+  def put(obj: DataObject): Reference
+}
