+++ transactor/src/test/scala/io/mediachain/transactor/DummyContext.scala
@@ -2,10 +2,11 @@ package io.mediachain.transactor
 
 import io.atomix.copycat.server.CopycatServer
 import io.atomix.catalyst.transport.Address
+import io.mediachain.types.TransactorTypes.JournalClient
 
 case class DummyContext(
   server: CopycatServer, 
-  client: Types.JournalClient, 
+  client: JournalClient,
   store: Dummies.DummyStore,
   logdir: String
 )
