+++ transactor/src/test/scala/io/mediachain/transactor/JournalCommitSpec.scala
@@ -5,7 +5,8 @@ import org.specs2.specification.{AfterAll, BeforeAll}
 import java.util.concurrent.{BlockingQueue, LinkedBlockingQueue, TimeUnit}
 import scala.concurrent.Await
 import scala.concurrent.duration.Duration
-import Types._
+import io.mediachain.types.DatastoreTypes._
+import io.mediachain.types.TransactorTypes._
 
 object JournalCommitSpec extends io.mediachain.BaseSpec
   with BeforeAll
